// ph-healthcare-frontent/src/generated/prisma/enums.ts
export const RoleKey = {
  SUPERADMIN: 'SUPERADMIN',
  ADMIN: 'ADMIN',
  DOCTOR: 'DOCTOR',
  PATIENT: 'PATIENT'
} as const

export type RoleKey = (typeof RoleKey)[keyof typeof RoleKey]
// ...
export const AppointmentStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CHECKED_IN: 'CHECKED_IN',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  NO_SHOW: 'NO_SHOW',
  RESCHEDULED: 'RESCHEDULED'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]


export const AppointmentType = {
  IN_PERSON: 'IN_PERSON',
  ONLINE: 'ONLINE',
  FOLLOW_UP: 'FOLLOW_UP',
  EMERGENCY: 'EMERGENCY'
} as const

export type AppointmentType = (typeof AppointmentType)[keyof typeof AppointmentType]